Make `list_files` private
This function doesn't need to be in the public API, which is used
by a program sending a command to roxy.
